src/email_service/requirements.txt
python-dotenv==1.1.1
fastapi==0.100.0
uvicorn[standard]==0.22.0